Refactor project index caching modules
