+++ src/plugin/src/project-index/constants.js
@@ -1,0 +1,1 @@
+export const PROJECT_MANIFEST_EXTENSION = ".yyp";
